refactor: Move article rendering into cached function on server side
